Prevent sometimes incorrect assignment of dtype to name with fromhdf5
Using the values in tables dtype fields dictionary-like object results
in a sometimes mismatch between the ctable's column name and expected
dtype, because the ordering is not guaranteed.

Instead, look up dtype for each column name in the same ordering that
the names that are passed to the ctable initialization.