introduce chunk._getitem_r(...), a thread-safe version of chunk._getitem(...)
`chunk._getitem(...)` depends on and modifies instance attributes of `chunk`.
The consequence is that in multi-threaded applications access to the `chunk`
instance has to be controlled by locking which prevent multi-threaded code
based on bcolz from scaling appropriately with the number of processors.

The introduced function `chunk._getitem_r(...)` provides the same functionality
without requiring the GIL or any locking to protect instance attributes. This
allows writing efficient multi-threaded cython algorithms based on bcolz.