Fix repr test on Windows.
The Problem is that carray.len for size 0 will return a long on windows but an
int on linux and hence this breaks the repr test. Fixing by casting to int
before calling shape. Would be better to fix this upstream in Cython perhaps,
but ok.

See also: https://github.com/Blosc/bcolz/pull/177