fix memory leak in chunk.getudata
This potentially fixes a memory leak in the 'getudata' method of the 'chunk'
class.

If you take a closer look at the code, you will see that a buffer 'dest' is
allocated. Then, the data is decompressed into this buffer. Next, a Python
string object is created. However, 'PyBytes_FromStringAndSize' (formerly
'PyString_FromStringAndSize') actually makes a copy of the data in memory.

https://docs.python.org/3.0/c-api/bytes.html

This in itself is already a bit of an issue, since we are making an unnecessary
copy of the uncompressed data in memory, but more importantly, the 'dest'
buffer is never 'freed'. This results in a memory leak which, depending on your
access patterns for the 'chunk' may result in memory exhaustion.

To reproduce, run the following against code w/o this patch and observe how
much memory is taken by the IPython process, e.g. with 'htop':

```
In [2]: import bcolz

In [3]: a = bcolz.arange(1e6)

In [4]: c = a.chunks[0]

In [5]: %timeit u = c.getudata()
1000 loops, best of 3: 454 µs per loop

In [6]: %timeit u = c.getudata()
1000 loops, best of 3: 437 µs per loop

In [7]: %timeit u = c.getudata()
1000 loops, best of 3: 455 µs per loop

In [8]: %timeit u = c.getudata()
1000 loops, best of 3: 446 µs per loop
```

With every run of timeit you should see a significant increase in used memory.
If you are lucky, you might get your machine to thrash and lock up.

Now try running the same example with the patch. You should see barely any
increase in memory (and the code might actually run faster too).

Ideally, we want to avoid the copy, but avoiding the memory leak is a good
start.